Refactor edit client relationship manager form to use new layout
