+++ src/apps/investments/client/projects/team/state.js
@@ -1,5 +1,3 @@
-export const TASK_SAVE_CLIENT_RELATIONSHIP_MANAGER =
-  'TASK_SAVE_CLIENT_RELATIONSHIP_MANAGER'
 export const TASK_EDIT_PROJECT_TEAM_MEMBERS = 'TASK_EDIT_PROJECT_TEAM_MEMBERS'
 export const TASK_SAVE_INVESTMENT_PROJECT_MANAGERS =
   'TASK_SAVE_INVESTMENT_PROJECT_MANAGERS'
