+++ src/apps/investments/client/projects/team/tasks.js
@@ -15,16 +15,6 @@ export const updateTeamMembers = ({ teamMembers, id }) => {
   })
 }
 
-export const saveClientRelationshipManager = ({
-  id,
-  clientRelationshipManagerId,
-}) =>
-  apiProxyAxios
-    .patch(`/v3/investment/${id}`, {
-      client_relationship_manager: clientRelationshipManagerId,
-    })
-    .then((results) => results.client_relationship_manager)
-
 export const saveProjectManagementStaff = ({
   id,
   projectAssuranceAdviserId,
