+++ src/apps/investments/controllers/team/index.js
@@ -1,9 +1,7 @@
-const editClientRelationshipManagement = require('./edit-client-relationship-management')
 const editProjectManagement = require('./edit-project-management')
 const editTeamMembers = require('./edit-team-members')
 
 module.exports = {
-  editClientRelationshipManagement,
   editProjectManagement,
   editTeamMembers,
 }
