--- src/apps/investments/views/team/edit-client-relationship-management.njk
@@ -1,8 +1,0 @@
-{% extends "../template.njk" %}
-
-{% block main_grid_right_column %}
-    {% component 'react-slot', {
-        id: 'edit-client-relationship-management',
-        props: props
-    } %}
-{% endblock %}
